Consolidate synonym pipeline into words/synonyms.py
Promote src/words/synonyms.py from a prompt wrapper to the canonical
synonym-generation library. It now owns normalization, the LLM call,
DerivativeForm storage, and the synonym_scan operation log. The
workqueue handler and ŠERNAS agent both delegate here, removing two
verbatim copies of _normalize_generated_forms and SYNONYM_FORM_MAP.

Add `alternate_spellings` as a generated category, mapped to the
existing `synonym_spelling` grammatical form already recognized by the
Barsukas display layer (e.g. gray/grey, color/colour).

The library API takes DataSourceConfig rather than a LinguisticClient
instance to avoid pulling wordfreq.translation.client at module import
time (which forms a cycle through langtools -> words).
