+++ src/words/__init__.py
@@ -2,7 +2,17 @@
 
 """Shared word-level generation helpers used by benchmarks and app flows."""
 
-from words.synonyms import build_synonyms_prompt, query_synonyms
+from words.synonyms import (
+    SYNONYM_FORM_MAP,
+    SYNONYMS_JSON_SCHEMA,
+    build_synonyms_prompt,
+    generate_synonyms_for_lemma,
+    normalize_generated_forms,
+    query_synonyms,
+    query_synonyms_from_llm,
+    record_synonym_processing_metadata,
+    store_synonym_forms,
+)
 from words.word2vec import rebuild_embeddings, refresh_lemma_embedding, suggest_similar_lemmas
 from words.translation import (
     build_multi_target_translation_prompt,
@@ -18,8 +28,15 @@ from words.ipa_pronunciation import (
 )
 
 __all__ = [
+    "SYNONYM_FORM_MAP",
+    "SYNONYMS_JSON_SCHEMA",
     "build_synonyms_prompt",
+    "generate_synonyms_for_lemma",
+    "normalize_generated_forms",
     "query_synonyms",
+    "query_synonyms_from_llm",
+    "record_synonym_processing_metadata",
+    "store_synonym_forms",
     "refresh_lemma_embedding",
     "rebuild_embeddings",
     "suggest_similar_lemmas",
